restaurant_management/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
python-dateutil==2.8.2
xlsxwriter==3.1.4
pillow==10.0.1
django-filter==23.2